+++ Nos/Views/HomeFeedView.swift
@@ -9,9 +9,6 @@ import SwiftUI
 import CoreData
 
 struct HomeFeedView: View {
-    
-    @AppStorage("keyPair") private var keyPair: KeyPair?
-    
     @Environment(\.managedObjectContext) private var viewContext
     
     @EnvironmentObject private var relayService: RelayService
@@ -34,7 +31,7 @@ struct HomeFeedView: View {
         })
         .overlay(Group {
             if events.isEmpty {
-                Text("No Events Yet! Add a relay to get started")
+                Localized.noEvents.view
             }
         })
         .toolbar {
@@ -43,11 +40,11 @@ struct HomeFeedView: View {
                     isCreatingNewPost.toggle()
                 }
                 label: {
-                    Label("Add Item", systemImage: "plus")
+                    Label(Localized.noEvents.string, systemImage: "plus")
                 }
             }
         }
-        .navigationTitle("Home Feed")
+        .navigationTitle(Localized.homeFeed.string)
         .task {
             load()
         }