+++ Nos/Views/NewPostView.swift
@@ -10,8 +10,16 @@ import CoreData
 import SwiftUINavigation
 
 struct NewPostView: View {
-    
-    @AppStorage("keyPair") private var keyPair: KeyPair?
+    private var keyPair: KeyPair? {
+		if let privateKeyData = KeyChain.load(key: KeyChain.keychainPrivateKey) {
+			let hexString = String(decoding: privateKeyData, as: UTF8.self)
+			if let pair = KeyPair(privateKeyHex: hexString) {
+				return pair
+			}
+        }
+		print("Error loading key pair from keychain")
+		return nil
+    }
     
     @Environment(\.managedObjectContext) private var viewContext
     
@@ -29,17 +37,17 @@ struct NewPostView: View {
                 TextEditor(text: $postText)
                     .frame(idealHeight: 180)
                 Button(action: publishPost) {
-                    Text("Publish")
+                    Localized.publish.view
                 }
             }
-            .navigationTitle("New Post")
+            .navigationTitle(Localized.newNote.string)
             .toolbar(content: {
                 ToolbarItem {
                     Button {
                         isPresented = false
                     }
                     label: {
-                        Text("Cancel")
+                        Localized.cancel.view
                     }
                 }
             })
@@ -50,9 +58,9 @@ struct NewPostView: View {
     private func publishPost() {
         guard let keyPair else {
             alert = AlertState(title: {
-                TextState("Error")
+                TextState(Localized.error.string)
             }, message: {
-                TextState("You need to enter a private key in Settings before you can publish posts.")
+                TextState(Localized.youNeedToEnterAPrivateKeyBeforePosting.string)
             })
             return
         }
@@ -74,7 +82,7 @@ struct NewPostView: View {
                 isPresented = false
             } catch {
                 alert = AlertState(title: {
-                    TextState("Error")
+                    TextState(Localized.error.string)
                 }, message: {
                     TextState(error.localizedDescription)
                 })